requirements.txt
streamlit
langchain-community
langchain-text-splitters
langchain-huggingface
langchain-chroma
sentence-transformers
chromadb
pypdf
python-dotenv
requests
huggingface-hub
